+++ projects/ui-framework/src/lib/table/table/table.enum.ts
@@ -1,0 +1,19 @@
+export enum TableType {
+  Primary = 'primary',
+  Secondary = 'secondary',
+}
+
+export enum SortDirections {
+  Asc = 'asc',
+  Desc = 'desc'
+}
+
+export enum RowSelection {
+  Multiple = 'multiple',
+  Single = 'single',
+}
+
+export enum PinDirection {
+  Left = 'left',
+  Right = 'right',
+}
