+++ projects/ui-framework/src/public_api.ts
@@ -289,6 +289,7 @@ export {
 // Table Module
 export { TableModule } from './lib/table/table.module';
 export * from './lib/table/table/table.interface';
+export * from './lib/table/table/table.enum';
 export { TableComponent } from './lib/table/table/table.component';
 export { AvatarCellComponent } from './lib/table/table-cell-components/avatar.component';
 export { ActionsCellComponent } from './lib/table/table-cell-components/actions-cell/actions-cell.component';
